# api/docs/swagger.yaml
basePath: /
definitions:
  handler.BedCreateRequest:
    properties:
      cell_number:
        example: 1
        type: integer
      is_locked:
        example: false
        type: boolean
    type: object
  handler.BedInitialRequest:
    properties:
      count:
        example: 5
        type: integer
    type: object
...
  model.Bed:
    properties:
      cell_number:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_locked:
        type: boolean
      user_id:
        type: integer
    type: object
  model.BedWithUserPlant:
    properties:
      cell_number:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_locked:
        type: boolean
      user_id:
        type: integer
      user_plant:
        $ref: '#/definitions/model.UserPlantWithSeed'
    type: object
...
  model.UserPlantWithSeed:
    properties:
      bed_id:
        type: integer
      created_at:
        type: string
      current_growth:
        type: integer
      gold_reward:
        type: integer
      growth_percent:
        type: integer
      id:
        type: integer
      seed_icon:
        type: string
      seed_id:
        type: integer
      seed_name:
        type: string
      target_growth:
        type: integer
      user_id:
        type: integer
      xp_reward:
        type: integer
    type: object
...
host: 10.155.36.40:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для FarmFocus - системы управления задачами и фермой
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FarmFocus API
  version: "1.0"
paths:
  /beds:
    get:
      consumes:
      - application/json
      description: Возвращает список всех грядок текущего пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Bed'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить все грядки пользователя
      tags:
      - beds
    post:
      consumes:
      - application/json
      description: Создает новую грядку для текущего пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Данные для создания грядки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BedCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Bed'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новую грядку
      tags:
      - beds
  /beds/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает грядку по указанному ID
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Bed ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Bed'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить грядку по ID
      tags:
      - beds
  /beds/{id}/lock:
    post:
      consumes:
      - application/json
      description: Блокирует указанную грядку
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Bed ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Заблокировать грядку
      tags:
      - beds
  /beds/{id}/unlock:
    post:
      consumes:
      - application/json
      description: Разблокирует указанную грядку
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Bed ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Разблокировать грядку
      tags:
      - beds
  /beds/available:
    get:
      consumes:
      - application/json
      description: Возвращает список разблокированных грядок пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Bed'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить доступные грядки
      tags:
      - beds
  /beds/cell/{cellNumber}:
    get:
      consumes:
      - application/json
      description: Возвращает грядку по номеру ячейки для текущего пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Cell Number
        in: path
        name: cellNumber
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Bed'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить грядку по номеру ячейки
      tags:
      - beds
  /beds/empty:
    get:
      consumes:
      - application/json
      description: Возвращает список пустых грядок пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Bed'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить пустые грядки
      tags:
      - beds
  /beds/init:
    post:
      consumes:
      - application/json
      description: Создает начальный набор грядок для пользователя
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Данные для создания начальных грядок
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BedInitialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать начальные грядки
      tags:
      - beds
  /beds/with-plants:
    get:
      consumes:
      - application/json
      description: Возвращает грядки с информацией о посаженных растениях
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.BedWithUserPlant'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить грядки с растениями
      tags:
      - beds
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"